helios-testing: Split TemporaryJobs
into Closeable and Rule implementations (implemented
by TemporaryJobsResource.
