+++ helios-testing/src/main/java/com/spotify/helios/testing/TemporaryJobs.java
@@ -17,14 +17,11 @@
 
 package com.spotify.helios.testing;
 
-import com.google.common.base.Throwables;
-import com.google.common.collect.Lists;
-import com.google.common.util.concurrent.MoreExecutors;
-import com.google.common.util.concurrent.ThreadFactoryBuilder;
-
 import com.spotify.helios.client.HeliosClient;
 import com.spotify.helios.common.Json;
 import com.spotify.helios.common.descriptors.Job;
+
+import com.google.common.collect.Lists;
 import com.typesafe.config.Config;
 import com.typesafe.config.ConfigFactory;
 import com.typesafe.config.ConfigList;
@@ -32,13 +29,10 @@ import com.typesafe.config.ConfigValue;
 import com.typesafe.config.ConfigValueFactory;
 import com.typesafe.config.ConfigValueType;
 
-import org.junit.rules.TestRule;
-import org.junit.runner.Description;
-import org.junit.runners.model.MultipleFailureException;
-import org.junit.runners.model.Statement;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.Closeable;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
@@ -50,31 +44,22 @@ import java.util.Date;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
-import java.util.concurrent.Callable;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.Future;
 import java.util.concurrent.ThreadLocalRandom;
-import java.util.concurrent.ThreadPoolExecutor;
 
 import static com.google.common.base.Charsets.UTF_8;
 import static com.google.common.base.Optional.fromNullable;
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
 import static java.lang.Integer.toHexString;
-import static java.lang.String.format;
 import static java.util.concurrent.TimeUnit.MINUTES;
-import static java.util.concurrent.TimeUnit.SECONDS;
 
-public class TemporaryJobs implements TestRule {
+public class TemporaryJobs implements Closeable {
 
   private static final Logger log = LoggerFactory.getLogger(TemporaryJobs.class);
 
   private static final String HELIOS_TESTING_PROFILE = "helios.testing.profile";
   private static final String HELIOS_TESTING_PROFILES = "helios.testing.profiles.";
   private static final Prober DEFAULT_PROBER = new DefaultProber();
-  private static final long JOB_HEALTH_CHECK_INTERVAL_MILLIS = SECONDS.toMillis(5);
   private static final long DEFAULT_DEPLOY_TIMEOUT_MILLIS = MINUTES.toMillis(10);
 
   private final HeliosDeployment heliosDeployment;
@@ -87,14 +72,6 @@ public class TemporaryJobs implements TestRule {
   private final Deployer deployer;
   private final Undeployer undeployer;
   private final String jobPrefix;
-
-  private final ExecutorService executor = MoreExecutors.getExitingExecutorService(
-      (ThreadPoolExecutor) Executors.newFixedThreadPool(
-          1, new ThreadFactoryBuilder()
-              .setNameFormat("helios-test-runner-%d")
-              .setDaemon(true)
-              .build()),
-      0, SECONDS);
 
   private static volatile HeliosDeployment soloDeployment;
 
@@ -134,46 +111,6 @@ public class TemporaryJobs implements TestRule {
     }
 
     return soloDeployment;
-  }
-
-  /**
-   * Perform setup. This is normally called by JUnit when TemporaryJobs is used with @Rule.
-   * If @Rule cannot be used, call this method before calling {@link #job()}.
-   *
-   * Note: When not being used as a @Rule, jobs will not be monitored during test runs.
-   */
-  public void before() {
-    deployer.readyToDeploy();
-  }
-
-  /**
-   * Perform teardown. This is normally called by JUnit when TemporaryJobs is used with @Rule.
-   * If @Rule cannot be used, call this method after running tests.
-   */
-  public void after() {
-    // Stop the test runner thread
-    executor.shutdownNow();
-    try {
-      final boolean terminated = executor.awaitTermination(30, SECONDS);
-      if (!terminated) {
-        log.warn("Failed to stop test runner thread");
-      }
-    } catch (InterruptedException ignore) {
-    }
-
-    final List<AssertionError> errors = new ArrayList<>();
-
-    for (final TemporaryJob job : jobs) {
-      // Undeploying a job doesn't guarantee the container will be stopped immediately.
-      // Luckily TaskRunner tells docker to wait 120 seconds after stopping to kill the container.
-      // So containers that don't immediately stop **should** only stay around for at most 120
-      // seconds.
-      errors.addAll(undeploy(job.job(), job.hosts()));
-    }
-
-    for (final AssertionError error : errors) {
-      log.error(error.getMessage());
-    }
   }
 
   public TemporaryJobBuilder job() {
@@ -278,72 +215,20 @@ public class TemporaryJobs implements TestRule {
     return builder(profile).build();
   }
 
-  @Override
-  public Statement apply(final Statement base, final Description description) {
-    return new Statement() {
-      @Override
-      public void evaluate() throws Throwable {
-        before();
-        try {
-          perform(base);
-        } finally {
-          after();
-        }
-      }
-    };
-  }
-
-  private void perform(final Statement base) throws InterruptedException {
-    // Run the actual test on a thread
-    final Future<Object> future = executor.submit(new Callable<Object>() {
-      @Override
-      public Object call() throws Exception {
-        try {
-          base.evaluate();
-        } catch (MultipleFailureException e) {
-          // Log the stack trace for each exception in the MultipleFailureException, because
-          // stack traces won't be logged when this is caught and logged higher in the call stack.
-          final List<Throwable> failures = e.getFailures();
-          log.error(format("MultipleFailureException contains %d failures:", failures.size()));
-          for (int i = 0; i < failures.size(); i++) {
-            log.error(format("MultipleFailureException %d:", i), failures.get(i));
-          }
-          throw Throwables.propagate(e);
-        } catch (Throwable throwable) {
-          Throwables.propagateIfPossible(throwable, Exception.class);
-          throw Throwables.propagate(throwable);
-        }
-        return null;
-      }
-    });
-
-    // Monitor jobs while test is running
-    while (!future.isDone()) {
-      Thread.sleep(JOB_HEALTH_CHECK_INTERVAL_MILLIS);
-      verifyJobsHealthy();
-    }
-
-    // Rethrow test failure, if any
-    try {
-      future.get();
-    } catch (ExecutionException e) {
-      final Throwable cause = (e.getCause() == null) ? e : e.getCause();
-      throw Throwables.propagate(cause);
-    }
-  }
-
-  private void verifyJobsHealthy() throws AssertionError {
-    for (final TemporaryJob job : jobs) {
-      Jobs.verifyHealthy(job.job(), client);
-    }
-  }
-
   String jobPrefix() {
     return jobPrefix;
   }
 
   public HeliosClient client() {
     return client;
+  }
+
+  List<TemporaryJob> jobs() {
+    return jobs;
+  }
+
+  Deployer deployer() {
+    return deployer;
   }
 
   public List<AssertionError> undeploy(final Job job, final List<String> hosts) {
@@ -364,6 +249,23 @@ public class TemporaryJobs implements TestRule {
 
   static Builder builder(final String profile, final Map<String, String> env) {
     return new Builder(profile, HeliosConfig.loadConfig("helios-testing"), env);
+  }
+
+  @Override
+  public void close() {
+    final List<AssertionError> errors = new ArrayList<>();
+
+    for (final TemporaryJob job : jobs) {
+      // Undeploying a job doesn't guarantee the container will be stopped immediately.
+      // Luckily TaskRunner tells docker to wait 120 seconds after stopping to kill the container.
+      // So containers that don't immediately stop **should** only stay around for at most 120
+      // seconds.
+      errors.addAll(undeploy(job.job(), job.hosts()));
+    }
+
+    for (final AssertionError error : errors) {
+      log.error(error.getMessage());
+    }
   }
 
   public static class Builder {
